Adds statistical confidence to variations.
This commit adds statisitcal confidence to the variation grid through use
of Z-Tests and normal distribution curves.

We first calculate a Z value for each variation (compared to the control)
and use a normal distribution function (NORMSDIST) to generate a probability
for the variation's conversion rate. If the probability is less than 5%, we're
95% confident that the test has accurate results.

We've used the following function for our normal distribution, modified to
remove the $sign variable (which caused incorrectsingle-tail results):
http://www.codeproject.com/Articles/408214/Excel-Function-NORMSDIST-z

We use the following guide to create our Z-tests:
http://20bits.com/article/statistical-analysis-and-ab-testing

The statistical methods for these are inside a new `Statistics` helper.