Adds an 'Improved by' column to variations.
We've added a column which shows the % increase in conversion rates between
the each variation and the control. If there's a dcrease the text is shown in
red, and there is JS which highlights the most improved variation by adding a
`best` class to the table row (highlighting the row's text in green).