Fixes worse variant bugs & best variant highlight
There were a couple of bugs with the Confidence and Improvement columns which
didn't show a negative result - they calculated improvements as positive and
still calculated a confidence even though the variation was worse. These have
now been fixed.

We've also added JS to detect whether the best variation is statistically
significant. We only want to highlight a winning result's row if we're
statistically confident that it's the winner, so JS now detects this.

We've also added a background image to the winning variation's row.