Allows injection of XML into any frontend request.
This commit allows us to inject variation XML into any frontend request. We
only need to know the module, controller and action names in order to inject
variation XML into the page layout.

We've done this by using the `controller_action_layout_generate_xml_before`
event to load tests and variations along with the other event hooks.

We've also started to implement an Ajax request to the sales website, allowing
us to track each download and which websites it has been implemented on.